requirements.txt
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
Deprecated==1.2.13
distlib==0.3.4
filelock==3.7.1
idna==3.3
platformdirs==2.5.2
pycparser==2.21
PyGithub==1.55
PyJWT==2.4.0
PyNaCl==1.5.0
PyYAML==6.0
requests==2.28.0
six==1.16.0
urllib3==1.26.9
virtualenv==20.14.1
wrapt==1.14.1
yamale==4.0.4